seperating model files
