+++ traiders/backend/api/admin.py
@@ -1,11 +1,10 @@
 from django.contrib import admin
-from django.contrib.auth.admin import UserAdmin
+from .models.user import User
+from .models.equipment import Equipment
+from .models.investment import ManualInvestment
+from .models.parity import Parity
 
-from .models.models import Equipment, ManualInvestment, Parity
-from .models.models import User
-
-
-admin.site.register(User, UserAdmin)
+admin.site.register(User)
 admin.site.register(Equipment)
 admin.site.register(ManualInvestment)
 admin.site.register(Parity)
