+++ traiders/backend/api/models/__init__.py
@@ -1,2 +1,4 @@
-from .models import Equipment, ManualInvestment, Parity
-from .models import User
+from .equipment import Equipment
+from .investment import ManualInvestment
+from .parity import Parity
+from .user import User
